+++ src/common/utils/index.ts
@@ -3,3 +3,4 @@ export * from './scroll';
 export { default as debounce } from './debounce';
 export * from './typeCheck';
 export * from './unit';
+export * from './getTimeBound';
