+++ src/components/UI/organisms/LectureList/SearchedLectureList.tsx
@@ -8,6 +8,7 @@ import { isString } from '@/common/utils/typeCheck';
 import { useQuery, useReactiveVar } from '@apollo/client';
 import { LectureListSkeleton } from '@/components/Skeleton';
 import { LECTURE_INFOS } from '@/queries';
+import { getTimeBoundByDay } from '@/common/utils';
 
 const SearchedLectureList = () => {
   const { timeTableStore, snackbarStore, lectureInfoStore } = useStores();
@@ -35,15 +36,6 @@ const SearchedLectureList = () => {
 
   if (error) return <p>Error :(</p>;
 
-  const getTimeBoundByDay = (day: string) => {
-    if (day === '월') return { start: 0, end: 1440 };
-    if (day === '화') return { start: 1440, end: 2880 };
-    if (day === '수') return { start: 2880, end: 4320 };
-    if (day === '목') return { start: 4320, end: 5760 };
-    if (day === '금') return { start: 5760, end: 7200 };
-    if (day === '토') return { start: 7200, end: 8640 };
-    return { start: 0, end: 8640 };
-  };
   const getFilteredByDepartmentLectures = (lectures: LectureInfos[]) => {
     if (department) return lectures.filter((lecture: LectureInfos) => lecture.department === department);
     return lectures;
